[#175883778] store added satispay
